[Planner] Restructure features and models; add docs
