+++ NOTES.md
@@ -1,4 +1,5 @@
 # Planner
+- 2025-08-08 00:20 UTC: Restructured features and models into top-level packages; added project docs.
 - 2025-07-29 00:15 UTC: Reviewed design notes; created plan-2025-07-29.md summarizing remaining tasks T9,T10,T12,T16,T17.
 - 2025-07-31 00:15 UTC: Reviewed repository status; created plan-2025-07-31.md with new tasks T18-T20 and moved T15 to completed.
 - 2025-08-01 00:10 UTC: Created plan-2025-08-01.md with tasks T21-T25 added to open list.
