+++ WORKFLOW.md
@@ -1,0 +1,7 @@
+# Workflow
+
+1. Planner writes tasks in `TASKS.md` and design notes under `design/`.
+2. Coder implements code in `src/`, `features/` and `models/`.
+3. Synthesizer assembles playbooks and docs.
+4. Tester runs `black`, `flake8` and `pytest`.
+5. Reviewer performs final review and updates `CHANGELOG.md`.
