+++ features/__init__.py
@@ -1,0 +1,5 @@
+"""Feature engineering package."""
+
+from .pipeline import compute_features, from_db, run_pipeline
+
+__all__ = ["compute_features", "from_db", "run_pipeline"]
