+++ models/__init__.py
@@ -1,0 +1,5 @@
+"""Model training utilities."""
+
+from .train import train
+
+__all__ = ["train"]
