+++ src/trading_platform/features/__init__.py
@@ -1,5 +1,6 @@
-"""Feature engineering package."""
+"""Backwards compatibility wrapper for feature functions."""
 
-from .pipeline import compute_features, from_db, run_pipeline
+from features import pipeline
+from features.pipeline import compute_features, from_db, run_pipeline
 
-__all__ = ["compute_features", "from_db", "run_pipeline"]
+__all__ = ["compute_features", "from_db", "run_pipeline", "pipeline"]
